find stable solutions to xor challenge
previously only a fixed number of iterations were being performed when
testing networks for xor behavior fitness. this had the result of
favoring some recurrent networks that produce unstable output and only
solve the problem for a non-continuous set of iteration counts.

this commit tests networks at a range iterations to assure that
solutions are fit for all of them. this method is similar to waiting for
output to stablize before calculating fitness.

this change has the unfortunate short term effect of increasing the
number of generations and test cases required to find a proper solution.
experimentation with settings may improve performance.