added option for limiting archive size
in search spaces that allow for vast numbers of behaviors, calculating
behavior sparsities of a generation is very costly, as each behavior
must be compared to all behaviors in the current generation along with
all behaviors in the archive. setting an upper limit on the number of
behaviors to keep in the archive mitigates this problem. the archive's
membership is fifo.